Refactor pool program state mod into objects
