+++ client/src/network/instruction/pool_create.rs
@@ -13,11 +13,11 @@ pub fn pool_create(admin: Pubkey, name: String, size: usize) -> Instruction {
             AccountMeta::new(admin, true),
             AccountMeta::new_readonly(clockwork_network_program::objects::Config::pubkey(), false),
             AccountMeta::new(
-                clockwork_pool_program::state::Pool::pubkey(name.clone()),
+                clockwork_pool_program::objects::Pool::pubkey(name.clone()),
                 false,
             ),
             AccountMeta::new_readonly(clockwork_pool_program::ID, false),
-            AccountMeta::new_readonly(clockwork_pool_program::state::Config::pubkey(), false),
+            AccountMeta::new_readonly(clockwork_pool_program::objects::Config::pubkey(), false),
             AccountMeta::new(clockwork_network_program::objects::Rotator::pubkey(), false),
             AccountMeta::new_readonly(system_program::ID, false),
         ],
