+++ client/src/pool/mod.rs
@@ -1,4 +1,4 @@
 pub mod instruction;
 
-pub use clockwork_pool_program::state;
+pub use clockwork_pool_program::objects;
 pub use clockwork_pool_program::ID;
