--- programs/network/src/id.rs
@@ -1,3 +1,0 @@
-use anchor_lang::prelude::declare_id;
-
-declare_id!("7uTtwxSLrUs6rTWzqYnvoQ8JnmcWT9WPDdoN1ojmvD6t");
