--- programs/pool/src/id.rs
@@ -1,3 +1,0 @@
-use anchor_lang::prelude::declare_id;
-
-declare_id!("CZZNnqWESL8qvcYktJ244YZo79L12g7EpJf646NgnKkb");
