+++ programs/pool/src/instructions/config_update.rs
@@ -1,4 +1,4 @@
-use {crate::state::*, anchor_lang::prelude::*};
+use {crate::objects::*, anchor_lang::prelude::*};
 
 #[derive(Accounts)]
 #[instruction(settings: ConfigSettings)]
@@ -6,7 +6,11 @@ pub struct ConfigUpdate<'info> {
     #[account(mut)]
     pub admin: Signer<'info>,
 
-    #[account(mut, seeds = [SEED_CONFIG], bump, has_one = admin)]
+    #[account(
+        mut, 
+        address = Config::pubkey(),
+        has_one = admin
+    )]
     pub config: Account<'info, Config>,
 }
