+++ programs/pool/src/lib.rs
@@ -1,14 +1,13 @@
 pub mod errors;
-pub mod id;
-pub mod state;
+pub mod objects;
 
 mod instructions;
 
-pub use id::ID;
-
 use anchor_lang::prelude::*;
 use instructions::*;
-use state::*;
+use objects::*;
+
+declare_id!("CZZNnqWESL8qvcYktJ244YZo79L12g7EpJf646NgnKkb");
 
 #[program]
 pub mod pool_program {
